requirements.txt
# Core
numpy==1.26.4
pandas==2.2.2

# NetCDF + scientific
xarray==2024.3.0
netCDF4==1.7.1
scipy==1.13.1

# Visualization
matplotlib==3.9.0
seaborn==0.13.2
cartopy==0.23.0

# Jupyter
jupyterlab==4.2.1
notebook==7.2.1

# Utilities
tqdm==4.66.4

# Optional (interactive apps / dashboards)
plotly==5.22.0
streamlit==1.36.0

# env
python-dotenv

# scientific depth calculations
gsw==3.6.0

# CPU usage and for model creation
joblib==1.5.2